Refactor post rendering into PostCard component
